Perform routing between two screens
Code runs with errors from this commit on, but the error does not stop the app and I can't see any performance issues. The errors also do not stop the app in the deployed version. Hence, I am committing with error still in the code.

I can't find a way to fix it yet :(